# api/openapi.yaml
openapi: 3.1.0
info:
  title: BossTransfer API
  version: 2.1.0
  description: 中央授权、115 资源搜索、客户秒传与本地下载 API。
servers:
  - url: /api/v1
paths:
...
  /manager/licenses:
    get:
      summary: 客户授权列表
      responses:
        "200": { description: 授权列表 }
    post:
      summary: 创建客户授权并签发一次性授权码
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CreateLicense" }
      responses:
        "201": { description: 已创建，完整授权码只在本响应出现 }
  /manager/licenses/{licenseId}:
    put:
      summary: 修改有效期与设备数
      parameters:
        - $ref: "#/components/parameters/LicenseID"
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/UpdateLicense" }
      responses:
        "200": { description: 已更新 }
    delete:
      summary: 撤销客户授权
      parameters:
        - $ref: "#/components/parameters/LicenseID"
      responses:
        "200": { description: 已撤销 }
...
components:
  securitySchemes:
    managerSession:
      type: apiKey
      in: cookie
      name: bosstransfer_manager_session
    clientSession:
      type: apiKey
      in: cookie
      name: bosstransfer_client_session
    deviceBearer:
      type: http
      scheme: bearer
  parameters:
    LicenseID:
      in: path
      name: licenseId
      required: true
      schema: { type: string }
...
  schemas:
...
    CreateLicense:
      type: object
      required: [customer, device_limit]
      properties:
        customer: { type: string, maxLength: 256 }
        expires_at: { type: string, format: date-time }
        device_limit: { type: integer, minimum: 1, maximum: 10000 }
    UpdateLicense:
      type: object
      required: [device_limit]
      properties:
        expires_at: { type: string, format: date-time }
        device_limit: { type: integer, minimum: 1, maximum: 10000 }
...